+++ CHANGELOG.md
@@ -45,3 +45,7 @@ CHANGELOG
 
 1.0.6 (14/11/2021)
 - setup.py redefined with metadata.
+
+1.0.7 (20/11/2021)
+- unin package becomes unin library.
+- duality >= 3.3.6 library changes applied to unin library.